finished Item component
